Decloupe libompfile from libomp
This moves in the direction of making libompfile export all the user API
for files. At first I planned to integrate it with the OpenMP scheduler,
but I will leave this for later.

This will simplify the linking process, as only libompfile links to
libomp, and we avoid the circular problem of having libomp call
libompfile.