Refactor libompfile to avoid circular dependencies
To ease the development process, we split libompfile from libomp.
We do this because libomp will not link against libstdc++. We
must also avoid a circular linking process, so for now only
libompfile links to libomp, and not the other way around.

We also export the headers to the build folder so its easier
to iterate as we develop.

When we need to interact with the schduler and make libomp
guarantee the progress of asnc I/O, we can rely on dlopen
to get the symbols on the other way of the dependency links.