[feat] Add posix backend
This commit adds a POSIX backend to manipulate files. So far, it plays
really well with the existing interface, since we only suport Open,
Close, Write, Read, Seed, ReadAt and WriteAt.

I think that we may have trouble doing the work around for unsupported
actions. I still wonder if the workarounds should be kept in the backend
or the agnostic interface, with the fallback logic.

I will brainstorm this design when the issue arises.